[#11] chores: 코드 위치 일부 수정
